Renamed some packages/classes
