+++ org.metaborg.example.api.analysis/src/main/java/org/metaborg/example/api/analysis/Main.java
@@ -4,7 +4,7 @@ package org.metaborg.example.api.analysis;
 import java.io.IOException;
 
 import org.metaborg.core.MetaborgException;
-import org.metaborg.example.api.analysis.parser.Parser;
+import org.metaborg.example.api.analysis.frontend.Frontend;
 import org.metaborg.example.api.analysis.syntax.Root;
 import org.metaborg.spoofax.core.Spoofax;
 import org.metaborg.util.log.ILogger;
@@ -19,16 +19,16 @@ public class Main {
 		
 		try(Spoofax spoofax = new Spoofax()) {
 		
-			Parser parser = new Parser(spoofax, languageResource);
+			Frontend parser = new Frontend(spoofax, languageResource);
 			// initialize everything by parsing/analysing a dummy expression
-			Root root = parser.parse("42");
+			Root root = parser.analyze("42");
 
 			ILogger logger = LoggerUtils.logger(Main.class);
 			
 			logger.info("start parsing");
-			root = parser.parse("let x = 21 in x + let x = 7 in x * 3 end end");
+			root = parser.analyze("let x = 21 in x + let x = 7 in x * 3 end end");
 			logger.info(root.toString());
-			root = parser.parseFile(filePath);
+			root = parser.analyzeFile(filePath);
 			logger.info(root.toString());
 			
 		} catch (MetaborgException | IOException e) {
