+++ org.metaborg.example.api.analysis/src/main/java/org/metaborg/example/api/analysis/frontend/Factory.java
@@ -1,4 +1,4 @@
-package org.metaborg.example.api.analysis.parser;
+package org.metaborg.example.api.analysis.frontend;
 
 import org.metaborg.example.api.analysis.SpoofaxUtil;
 import org.metaborg.example.api.analysis.syntax.Add;
